+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 2.8
+
+- Add Certificate Transparency Submission
+- Small formatting changes
+
+
+## 2.7
+
+- Add debian weak keys check.
+- Add Lets Encrypt Signing certs: https://letsencrypt.org/2015/06/04/isrg-ca-certs.html
+- Check heartbeat extension.
+- Set title based on check.
+- Reset warning count in menu correctly.
+- Remove spaces from json variable names.
+
 ## 2.6
 
 - Fix testing of IPv6 only hosts.